Fix: 알림 채널 생성은 앱 시작 시 한 번만 생성하도록 변경
